+++ src/main/kotlin/infrastructure/aggregator/shared/ProviderCurrencyAdapter.kt
@@ -3,7 +3,6 @@ package infrastructure.aggregator.shared
 import application.port.outbound.CurrencyAdapter
 import shared.value.Currency
 import java.math.BigDecimal
-import java.math.BigInteger
 
 /**
  * Shared currency adapter wrapper for aggregators that use the standard conversion
@@ -12,11 +11,11 @@ import java.math.BigInteger
  * Used by Pateplay and Pragmatic aggregators.
  */
 class ProviderCurrencyAdapter(private val currencyAdapter: CurrencyAdapter) {
-    suspend fun convertSystemToProvider(amount: BigInteger, currency: Currency): BigDecimal {
+    suspend fun convertSystemToProvider(amount: Long, currency: Currency): BigDecimal {
         return currencyAdapter.convertFromSystem(amount, currency)
     }
 
-    suspend fun convertProviderToSystem(amount: BigDecimal, currency: Currency): BigInteger {
+    suspend fun convertProviderToSystem(amount: BigDecimal, currency: Currency): Long {
         return currencyAdapter.convertToSystem(amount, currency)
     }
 }
